+++ websourceblog.php
@@ -45,7 +45,7 @@ class WebsourceBlog extends Module
     {
         $this->name = 'websourceblog';
         $this->tab = 'front_office_features';
-        $this->version = '1.0.0';
+        $this->version = '1.0.1';
         $this->author = 'Websource';
         $this->need_instance = 0;
         $this->bootstrap = true;